app/tools/optimization_tools.py
from typing import Dict, Any
import pandas as pd
from skopt import gp_minimize
from skopt.space import Real, Integer
import numpy as np

def optimize_strategy_params(
    strategy_logic_description: str,
    data_dict: Dict,
    param_space: dict,
    method: str = 'bayesian'
) -> Dict[str, Any]:
    try:
        # Convert data to DataFrame
        df = pd.DataFrame(data_dict['data'])
        
        # Define parameter space for skopt
        dimensions = []
        param_names = []
        for name, space in param_space.items():
            param_names.append(name)
            if space['type'] == 'real':
                dimensions.append(Real(space['low'], space['high']))
            elif space['type'] == 'integer':
                dimensions.append(Integer(space['low'], space['high']))
        
        # Define objective function
        def objective(params):
            param_dict = dict(zip(param_names, params))
            try:
                # Execute strategy with parameters
                result = execute_strategy(
                    df,
                    strategy_logic_description,
                    param_dict
                )
                # Return negative Sharpe ratio (we minimize)
                return -result['sharpe_ratio']
            except Exception:
                return np.inf
        
        # Run optimization
        result = gp_minimize(
            objective,
            dimensions,
            n_calls=50,
            random_state=42
        )
        
        best_params = dict(zip(param_names, result.x))
        return {
            "success": True,
            "best_params": best_params,
            "optimization_score": -result.fun
        }
        
    except Exception as e:
        return {
            "success": False,
            "error": str(e)
        }

def execute_strategy(
    df: pd.DataFrame,
    logic_description: str,
    params: Dict[str, Any]
) -> Dict[str, float]:
    """Execute strategy based on description and parameters."""
    # This is a simplified example
    returns = pd.Series(0.0, index=df.index)
    
    if "moving_average" in logic_description.lower():
        fast_ma = df['close'].rolling(params['fast_period']).mean()
        slow_ma = df['close'].rolling(params['slow_period']).mean()
        position = (fast_ma > slow_ma).astype(float)
        returns = position.shift(1) * df['close'].pct_change()
    
    # Calculate metrics
    sharpe_ratio = np.sqrt(252) * returns.mean() / returns.std()
    
    return {
        "sharpe_ratio": sharpe_ratio,
        "total_return": (1 + returns).prod() - 1,
        "max_drawdown": (returns.cumsum() - returns.cumsum().cummax()).min()
    }