requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
structlog>=23.1.0
langchain>=0.0.300
langchain-google-genai>=0.0.5
autogen-agentchat~=0.2
langgraph>=0.0.10
vectorbt>=0.25.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
scikit-optimize>=0.9.0
llama-index>=0.8.0
alembic>=1.11.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.2

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
asyncpg>=0.28.0